feat: colocate setProfile util function with getProfile query
