+++ lib/profile/getProfile.ts
@@ -1,9 +1,9 @@
 import { gql } from "../helpers/gql";
 import { TransactionEdge } from "arweave-graphql";
-import { GetUseProfileProps, Profile } from "@/types";
-import { setProfileInfo } from "./setProfileInfo";
+import { GetUserProfileProps, Profile } from "@/types";
+import { gateway } from "@/utils";
 
-export const getProfile = async ({ address }: GetUseProfileProps) => {
+export const getProfile = async ({ address }: GetUserProfileProps) => {
   if (!address) {
     throw new Error("No address has been given.");
   }
@@ -29,4 +29,36 @@ export const getProfile = async ({ address }: GetUseProfileProps) => {
     console.error(error);
     throw new Error(error.message);
   }
+};
+
+const setProfileInfo = (edge: TransactionEdge): Profile => {
+  // casting as the filter in query func is/should be ensuring value exists
+  const name = edge.node.tags.find((x) => x.name === "Name")?.value as string;
+
+  const handle = edge.node.tags.find((x) => x.name === "Handle")?.value;
+  const bio = edge.node.tags.find((x) => x.name === "Bio")?.value;
+
+  const thumbnailId = edge.node.tags.find((x) => x.name === "Thumbnail")?.value;
+  const avatarId = edge.node.tags.find((x) => x.name === "Avatar")?.value;
+  const bannerId = edge.node.tags.find((x) => x.name === "Banner")?.value;
+
+  const thumbnailSrc = thumbnailId ? gateway() + "/" + thumbnailId : undefined;
+  const avatarSrc = avatarId ? gateway() + "/" + avatarId : undefined;
+  const bannerSrc = bannerId ? gateway() + "/" + bannerId : undefined;
+
+  const txid = edge.node.id;
+  const cursor = edge.cursor;
+  const owner = edge.node.owner.address;
+
+  return {
+    txid,
+    addr: owner,
+    name,
+    handle,
+    bio,
+    thumbnailSrc,
+    avatarSrc,
+    bannerSrc,
+    cursor,
+  };
 };
